Moving Settings logic to /Domain/Settings
