+++ src/Backend/Modules/Mailmotor/Actions/Ping.php
@@ -11,8 +11,8 @@ namespace Backend\Modules\Mailmotor\Actions;
 
 use Backend\Core\Engine\Base\ActionIndex;
 use Backend\Core\Engine\Model;
-use Backend\Modules\Mailmotor\Command\SaveSettings;
-use Backend\Modules\Mailmotor\Event\SettingsSavedEvent;
+use Backend\Modules\Mailmotor\Domain\Settings\Command\SaveSettings;
+use Backend\Modules\Mailmotor\Domain\Settings\Event\SettingsSavedEvent;
 
 /**
  * This tests the api
