+++ src/Backend/Modules/Mailmotor/Resources/config/forms.yml
@@ -1,6 +1,6 @@
 services:
   mailmotor.form.settings:
-    class: Backend\Modules\Mailmotor\Form\SettingsType
+    class: Backend\Modules\Mailmotor\Domain\Settings\SettingsType
     arguments:
       - "@mailmotor.manager.subscriber_gateway"
     tags:
